fix: 修改 AIToDoListDetail 组件
