+++ app/renderer/src/main/src/pages/ai-re-act/aiReActChat/aiToDoList/type.ts
@@ -7,6 +7,9 @@ export interface AIToDoListProps {
   /** 禁止展开收起 */
   bannedExpand?: boolean
 }
+export interface AIToDoListDetailProps {
+  todoData: TodoListCardData
+}
 export interface AIToDoListItemProps {
   item: AIAgentGrpcApi.TodoListUpdateItem
 }
